Initialize the theme on client-rendered routes
Not-found pages under the dynamic section routes are rendered entirely on
the client (the PPR fallback shell is empty), and React doesn't execute
inline scripts it inserts. So the theme/uwu/platform init never ran there:
the page came up light regardless of preference, and the theme toggle threw
because window.__setPreferredTheme didn't exist.

Move the script to a shared module and evaluate it from a client effect when
the inline version hasn't run.
