+++ src/app/layout.tsx
@@ -8,7 +8,12 @@
 import type {Metadata, Viewport} from 'next';
 import Script from 'next/script';
 import {siteConfig} from '../siteConfig';
-import {AnalyticsTracker, ScrollRestoration} from './clientEffects';
+import {
+  AnalyticsTracker,
+  ScrollRestoration,
+  ThemeInitFallback,
+} from './clientEffects';
+import {themeScript} from './themeScript';
 
 import '@docsearch/css';
 import '../styles/algolia.css';
@@ -47,104 +52,6 @@ export const metadata: Metadata = {
     'google-site-verification': 'sIlAGs48RulR4DdP95YSWNKZIEtCqQmRjzn-Zq-CcD0',
   },
 };
-
-const themeScript = `
-(function () {
-  try {
-    let logShown = false;
-    function setUwu(isUwu) {
-      try {
-        if (isUwu) {
-          localStorage.setItem('uwu', true);
-          document.documentElement.classList.add('uwu');
-          if (!logShown) {
-            console.log('uwu mode! turn off with ?uwu=0');
-            console.log('logo credit to @sawaratsuki1004 via https://github.com/SAWARATSUKI/KawaiiLogos');
-            logShown = true;
-          }
-        } else {
-          localStorage.removeItem('uwu');
-          document.documentElement.classList.remove('uwu');
-          console.log('uwu mode off. turn on with ?uwu');
-        }
-      } catch (err) { }
-    }
-    window.__setUwu = setUwu;
-    function checkQueryParam() {
-      const params = new URLSearchParams(window.location.search);
-      const value = params.get('uwu');
-      switch(value) {
-        case '':
-        case 'true':
-        case '1':
-          return true;
-        case 'false':
-        case '0':
-          return false;
-        default:
-          return null;
-      }
-    }
-    function checkLocalStorage() {
-      try {
-        return localStorage.getItem('uwu') === 'true';
-      } catch (err) {
-        return false;
-      }
-    }
-    const uwuQueryParam = checkQueryParam();
-    if (uwuQueryParam != null) {
-      setUwu(uwuQueryParam);
-    } else if (checkLocalStorage()) {
-      document.documentElement.classList.add('uwu');
-    }
-  } catch (err) { }
-})();
-
-(function () {
-  function setTheme(newTheme) {
-    window.__theme = newTheme;
-    if (newTheme === 'dark') {
-      document.documentElement.classList.add('dark');
-    } else if (newTheme === 'light') {
-      document.documentElement.classList.remove('dark');
-    }
-  }
-
-  var preferredTheme;
-  try {
-    preferredTheme = localStorage.getItem('theme');
-  } catch (err) { }
-
-  window.__setPreferredTheme = function(newTheme) {
-    preferredTheme = newTheme;
-    setTheme(newTheme);
-    try {
-      localStorage.setItem('theme', newTheme);
-    } catch (err) { }
-  };
-
-  var initialTheme = preferredTheme;
-  var darkQuery = window.matchMedia('(prefers-color-scheme: dark)');
-
-  if (!initialTheme) {
-    initialTheme = darkQuery.matches ? 'dark' : 'light';
-  }
-  setTheme(initialTheme);
-
-  darkQuery.addEventListener('change', function (e) {
-    if (!preferredTheme) {
-      setTheme(e.matches ? 'dark' : 'light');
-    }
-  });
-
-  document.documentElement.classList.add(
-      window.navigator.platform.includes('Mac')
-      ? "platform-mac"
-      : "platform-win"
-  );
-})();
-`;
 
 // Translation forks load the fonts from react.dev too.
 const FONT_ORIGIN = 'https://react.dev';
@@ -214,6 +121,7 @@ export default function RootLayout({children}: {children: React.ReactNode}) {
         )}
       </head>
       <body className="font-text font-medium antialiased text-lg bg-wash dark:bg-wash-dark text-secondary dark:text-secondary-dark leading-base">
+        <ThemeInitFallback />
         <ScrollRestoration />
         <AnalyticsTracker />
         {children}
